+++ CHANGELOG.md
@@ -4,3 +4,10 @@
 - Generate Polygon EVM wallet (address, private key, mnemonic).
 - Register wallet with https://ga-api.onchainlabs.ch.
 - Store mnemonic in secure storage on device.
+
+## 2.0.0
+
+- Add Mint function
+- Add Minter admin wallet within the test app
+- Add Transfer function
+- Add Approve function